make the default protocol binding HTTP-POST
Per @puiterwijk in issue #32, ProtocolBinding
refers to the *response* not the request. In a 
previous commit I conflated the the two.

Because HTTP-POST is a sane generally correct thing to 
specify, and because errors occur when it is unspecified
we make it the default protocol binding.

This commit also adds a note (direct from the standard)
about the meaning of the field, to remind ourselves.