BREAKING: update schema to more accurately reflect the standard
This commit adds missing fields and structures. 

This commit turns some fields from pointers to a single optional struct into 
the more correct slice of struct. In those cases the field name also changes 
to plural.

This commit also improves the way XML signatures are modeled. It turns out that
some libraries (and maybe the actual SAML standard) require that elements appear
in a particular order, including <ds:Signature> elements. To model this we add
Signature fields as an *etree.Element. These structs also grow an Element() 
method that serializes the struct into a *etree.Element. This is required to 
correctly compute digests of the XML as specified by xmldsig.