Use RS256 rather than HS256
Using HS256 with the RSA private key (serialised as ASN.1 according
to PKCS1) as the secret is unconventional and perhaps dangerous. The
string of bytes isn't entirely random given that it starts with the
version number. The encoded RSA key will be larger than the block size
of the underlying hash and so to create a secret, the RSA key will be
hashed.

A more important problem is that we should be able to validate a jwt
issued by the middleware with just the RSA public key. By moving to
RS256, we can validate jwts without sharing the private key.

Fix randomBytes: should ensure that the entire buffer is full.