Add support for hardware security module (HSM) signing.
goxmldsig was recently updated to support HSMs. This is done by supporting
signing using the `crypto.Signer` directly, which is how HSMs are able to
sign things in golang.

This commit adds a new field `Signer` to the `IdentityProvider` which takes
precedence over the `Key` field when creating the `SigningContext`. The `Key`
field has been left in place to maintain backwards compatibility.

Tests have been updated to verify that this functions, and the `samlidp`
example identity provider has been updated such that it now supports a
`Signer` option. Additionally, `NewIdentifyProviderTest` has been corrected to
`NewIdentityProviderTest`.