Fixed samlsp to support FQDN as cookie identifiers
I didn't know that samlsp didn't support FQDN as cookie parameters.
If we passed in, for example, https://idp.apple.com, we'd only get back
"https" which wouldn't set the cookie for the actual domain,
idp.apple.com.

This phenomenon was because we used net.SplitHostPort which does not
like taking the scheme into consideration. With this, we are now able to
do just that while maintaining the previous implementation using
url.Parse and setting the domain to the Hostname.