Address GO-2024-3250 vulnerability
https://pkg.go.dev/vuln/GO-2024-3250

Addressed by bumping the github.com/golang-jwt/jwt/v4 library from
v4.5.0 to v4.5.1.

Based on a quick look at the calls to ParseWithClaims (the affected
function) I don't think there was necessarily a problem to begin
with (we treated all errors as fatal), but this will make `go vuln`
happy which is important.