Middleware.RequireAccount: handle error before reading session
m.RequireAccount checks if session is nil and only handles error when it
is. This is not idiomatic. Also, ErrNoSession is handled with direct
equality comparison. So, handle errors and protect against nil Session
before adding it to the request context. And when checking for
ErrNoSession, do so with errors.Is. This way, custom implementations of
SessionProvider can wrap ErrNoSession, if they so choose.